+++ src/hooks/business/useBadgeListViewModel.ts
@@ -14,6 +14,7 @@ export const useBadgeListViewModel = () => {
     return { isTagLoading, tagData: null };
   }
   console.log(tagData);
+  console.log(CourseBadges);
   console.log(DestinationBadges); //성공
 
   return { isTagLoading, tagData, DestinationBadges, CourseBadges, setCourseBadges, setDestinationBadges };
